vkd3d: Make enqueued_fence_count a uint32_t
Fixes atomic usage with it.

+++ libs/vkd3d/vkd3d_private.h
@@ -206,7 +206,7 @@ struct vkd3d_fence_worker
     bool should_exit;
     bool pending_fence_destruction;
 
-    size_t enqueued_fence_count;
+    uint32_t enqueued_fence_count;
     struct vkd3d_enqueued_fence
     {
         VkSemaphore vk_semaphore;
